requirements.txt
vkbottle==4.3.12
python-dotenv==1.0.1
SQLAlchemy==2.0.25
aiosqlite==0.19.0
